requirements.txt
# Invoice Physics - Python Dependencies
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database & Caching
redis>=5.0.0
neo4j>=5.14.0

# Graph Processing
networkx>=3.2.0
python-louvain>=0.16

# Fingerprinting & Similarity
datasketch>=1.6.0
sentence-transformers>=2.2.2
torch>=2.1.0

# LLM (Optional - for explanations)
llama-cpp-python>=0.2.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpie>=3.2.0

# Note: PyTorch Geometric (optional for GNN features)
# torch-geometric>=2.4.0

# File Upload Extraction
PyPDF2>=3.0.0
python-multipart>=0.0.9

# Gemini Agent Framework
google-genai>=0.3.0
